# docs/swagger.yaml
definitions:
...
  controller.PresignProfilePhotoGetResponse:
    properties:
      url:
        type: string
    type: object
...
  controller.UploadProfilePhotoResponse:
    properties:
      message:
        example: profile picture uploaded
        type: string
      public_url:
        example: http://localhost:9000/user-media/users/<sub>/profile.jpg
        type: string
    type: object
...
  controller.UserPublicInfoResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
    type: object
...
  models.User:
    properties:
      biography:
        type: string
      country:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_blocked:
        type: boolean
      job_function:
        type: string
      keycloak_id:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      phone_number_visible:
        type: boolean
      profile_photo_url:
        type: string
      sector:
        type: string
    type: object
  models.Badge:
    type: object
    properties:
      key:
        type: string
        example: profile_complete
      name:
        type: string
        example: Complete Profile
      description:
        type: string
        example: User completed all required profile information
  models.UserBadge:
    type: object
    properties:
      id:
        type: integer
        example: 1
      user_id:
        type: string
        format: uuid
      badge_key:
        type: string
        example: profile_complete
      earned_at:
        type: string
        format: date-time
      badge:
        $ref: '#/definitions/models.Badge'
info:
  contact: {}
paths:
...
  /users/{email}:
    get:
      description: Internal endpoint - requires X-Service-Token
      parameters:
      - description: Service token
        in: header
        name: X-Service-Token
        required: true
        type: string
      - description: User email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user by email (internal)
      tags:
      - Users
  /users/{firstname}/{lastname}:
    get:
      consumes:
      - application/json
      description: Returns only first_name, last_name and email for the matched user.
        Requires Bearer access token.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: First name
        in: path
        name: firstname
        required: true
        type: string
      - description: Last name
        in: path
        name: lastname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.UserPublicInfoResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user email by firstname-lastname
      tags:
      - Users
  /users/keycloak/{sub}:
    get:
      description: Get a single user by their Keycloak subject (sub) ID
      parameters:
      - description: Keycloak subject ID
        in: path
        name: sub
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Missing Keycloak sub
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user by Keycloak subject
      tags:
      - Users
  /users/me:
    put:
      consumes:
      - application/json
      description: Update logged-in user's profile data
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Updated user fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update my user profile
      tags:
      - Users
...
  /users/me/profile-photo:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a profile photo (multipart/form-data) to MinIO/S3, stores
        the public URL in the user profile, and returns the public URL.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Profile photo file (jpeg/png/webp). Max 5MB.
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.UploadProfilePhotoResponse'
        "400":
          description: Missing file / unsupported type / too large
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload profile photo
      tags:
      - Users
  /users/me/profile-photo/url:
    get:
      description: Returns a presigned GET URL to download the logged-in user's profile
        photo from MinIO/S3.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.PresignProfilePhotoGetResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User or profile photo not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Presign profile photo download
      tags:
      - Users
...
  /users/me/badges:
    get:
      description: Returns all badges earned by the authenticated user.
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer access token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/models.UserBadge'
        "401":
          description: Unauthorized
        "500":
          description: Could not load badges
      summary: Get my badges
      tags:
      - Badges
  /users/id/{id}/badges:
    get:
      description: Returns all badges earned by a user ID.
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
      - in: path
        name: id
        required: true
        type: string
        description: User ID (UUID)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/models.UserBadge'
        "400":
          description: Invalid user ID
        "401":
          description: Unauthorized
        "404":
          description: User not found
        "500":
          description: Could not load badges
      summary: Get badges by user ID
      tags:
      - Badges
...
swagger: "2.0"